backend/requirements.txt
flask
flask-cors
flask-socketio
eventlet
mysql-connector-python
requests
